requirements.txt
Kivy~=2.3.0
matplotlib~=3.9.2
backtrader~=1.9.78.123
yfinance~=0.2.50
pandas~=2.2.3
python-dotenv~=1.0.1
websocket-client~=1.8.0
numpy~=2.1.3